grpc: read the service config channel once instead of twice
